Fixes problem with reading EDN files in min build
This is an interesting bug. Having an extension of .edn makes figwheel's
server send out a "parsed" response instead of a plain string.  When
serving the site off a "normal" server, this doesn't happen and the data
read was failing.

This commit changes the extension of the data file to fix this problem,
and always parses the response on the client side.